Rename file (NBitcoin/UInt2561.cs -> NBitcoin/UInt256.cs)
